Ajustes na Listagem de Frequencias
Ajustes em Tipografia, Cores e Refatorações na estrutura da TableView
