--- YogaZone/YogaZone/Meditacao/MeditacaoHome/Components/AudioCell.swift
@@ -1,17 +1,0 @@
-import UIKit
-
-class AudioCell: UITableViewCell {
-
-    @IBOutlet weak var titleLabel: UILabel!
-    
-    @IBOutlet weak var selectButtonImage: UIImageView!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-    }
-    
-}
