+++ YogaZone/YogaZone/Meditacao/MeditacaoHome/MeditacaoHomeViewController.swift
@@ -1,7 +1,5 @@
 import UIKit
 
-private let reuseIdentifier = "AudioCell"
-
 class MeditacaoHomeViewController: UIViewController {
     
     @IBOutlet weak var meditationTableView: UITableView!
@@ -10,8 +8,6 @@ class MeditacaoHomeViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.meditationTableView.delegate = self
-        self.meditationTableView.dataSource = self
         
         setupData()
         setupUI()
@@ -31,10 +27,11 @@ extension MeditacaoHomeViewController {
 // MARK: UI Functions
 extension MeditacaoHomeViewController {
     func setupUI() {
-        
-        let nibCell = UINib(nibName: reuseIdentifier, bundle: nil)
-        self.meditationTableView.register(nibCell, forCellReuseIdentifier: reuseIdentifier)
-        self.meditationTableView.separatorColor = .clear
+        self.meditationTableView.delegate = self
+        self.meditationTableView.dataSource = self
+        self.meditationTableView.register(AudioCell.getNib(), forCellReuseIdentifier: AudioCell.identifier)
+        self.meditationTableView.separatorStyle = .none
+        self.meditationTableView.showsVerticalScrollIndicator = false
     }
 }
 
@@ -50,27 +47,23 @@ extension MeditacaoHomeViewController: UITableViewDelegate, UITableViewDataSourc
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = meditationTableView.dequeueReusableCell(withIdentifier: AudioCell.identifier, for: indexPath) as? AudioCell
+        
         let meditationFile = meditationFiles[indexPath.row]
+        cell?.setupCell(meditationFile: meditationFile)
         
-        let cell = meditationTableView.dequeueReusableCell(withIdentifier: reuseIdentifier, for: indexPath) as? AudioCell
-        cell?.titleLabel.text = meditationFile.title
-        cell?.selectButtonImage.image = #imageLiteral(resourceName: "medicacao-select")
-        
-        cell?.layer.cornerRadius = 8
-        cell?.layer.masksToBounds = true
         cell?.selectionStyle = .none
-        cell?.contentView.backgroundColor = UIColor(red: 113, green: 95, blue: 169, alpha: 0.8)
         
         return cell ?? UITableViewCell()
     }
     
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 50
+        return 69
     }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let selectedFile = meditationFiles[indexPath.row]
-        let vc = PracticeViewController(nibName: "PracticeViewController", bundle: nil)
+        let vc = PracticeViewController()
         
         vc.selectedFile = selectedFile
